Add phfs files endpoint support view and download
