+++ packages/graphql-server/src/app.ts
@@ -12,7 +12,6 @@ import morgan from 'koa-morgan';
 import * as GraphQLJSON from 'graphql-type-json';
 import { makeExecutableSchema } from 'graphql-tools';
 import { applyMiddleware } from 'graphql-middleware';
-import mime from 'mime';
 import url from 'url';
 import CrdClient, { InstanceTypeSpec, ImageSpec } from './crdClient/crdClientImpl';
 import * as system from './resolvers/system';
@@ -68,6 +67,7 @@ import Token from './oidc/token';
 import ApiTokenCache from './oidc/apiTokenCache';
 import { createMinioClient } from './utils/minioClient';
 import { mountTusCtrl } from './controllers/tusCtrl';
+import { mountStoreCtrl } from './controllers/storeCtrl';
 import { Telemetry } from './utils/telemetry';
 import { createDefaultTraitMiddleware } from './utils/telemetryTraits';
 import { isGroupBelongUser } from './utils/groupCheck';
@@ -614,42 +614,8 @@ export const createApp = async (): Promise<{app: Koa, server: ApolloServer, conf
   });
 
   if (config.enableStore) {
-    // phfs file download api
-    rootRouter.get('/files/(.*)', authenticateMiddleware, checkUserGroup,
-      async ctx => {
-        const objectPath = decodeURIComponent(ctx.request.path.split('/groups').pop());
-        let req;
-        try {
-          req = await minioClient.getObject(storeBucket, `groups${objectPath}`);
-        } catch (error) {
-          if (error.code === 'NoSuchKey') {
-            return ctx.status = 404;
-          } else {
-            logger.error({
-              component: logger.components.internal,
-              type: 'MINIO_GET_OBJECT_ERROR',
-              code: error.code,
-              message: error.message
-            });
-            ctx.res.end();
-          }
-        }
-
-        req.on('error', err => {
-          logger.error({
-            component: logger.components.internal,
-            type: 'MINIO_GET_OBJECT_ERROR',
-            message: err.message
-          });
-          ctx.res.end();
-        });
-
-        ctx.body = req;
-
-        const mimetype = mime.getType(objectPath);
-        ctx.set('Content-type', mimetype);
-      }
-    );
+    // store file download api
+    mountStoreCtrl(rootRouter, authenticateMiddleware, checkUserGroup, minioClient, storeBucket);
 
     // shared space proxy to tusd
     const tusProxyPath = `${staticPath}tus`;
